Separate graphql server and real-time ingesting
